[tests] Mv attempts tests in own file
