+++ tests/test_plugin/test_caused_dut_falure.py
@@ -101,75 +101,3 @@ def test_caused_dut_failure_id_first(pytester: Pytester, hardpy_opts: list):
     )
     result = pytester.runpytest(*hardpy_opts)
     result.assert_outcomes(failed=2)
-
-
-def test_caused_dut_failure_id_second_with_attempt(
-    pytester: Pytester,
-    hardpy_opts: list,
-):
-    """Caused DUT failure id is not filled if there is a successful attempt."""
-    pytester.makeconftest(
-        f"""
-        {import_header}
-
-        def finish_executing():
-            report = hardpy.get_current_report()
-
-            caused_dut_failure_id = report.caused_dut_failure_id
-            assert caused_dut_failure_id == "test_1::test_b"
-
-        {conftest_actions_after}
-    """,
-    )
-    pytester.makepyfile(
-        test_1="""
-        import pytest
-        import hardpy
-
-        @pytest.mark.attempt(2)
-        def test_a():
-            attempt = hardpy.get_current_attempt()
-            if attempt == 2:
-                return
-            assert False
-
-        def test_b():
-            assert False
-    """,
-    )
-    result = pytester.runpytest(*hardpy_opts)
-    result.assert_outcomes(passed=1, failed=1)
-
-def test_caused_dut_failure_id_second_with_critical(
-    pytester: Pytester,
-    hardpy_opts: list,
-):
-    """Caused DUT failure id is filled by critial test."""
-    pytester.makeconftest(
-        f"""
-        {import_header}
-
-        def finish_executing():
-            report = hardpy.get_current_report()
-
-            caused_dut_failure_id = report.caused_dut_failure_id
-            assert caused_dut_failure_id == "test_1::test_a"
-
-        {conftest_actions_after}
-    """,
-    )
-    pytester.makepyfile(
-        test_1="""
-        import pytest
-        import hardpy
-
-        @pytest.mark.critical
-        def test_a():
-            assert False
-
-        def test_b():
-            assert False
-    """,
-    )
-    result = pytester.runpytest(*hardpy_opts)
-    result.assert_outcomes(failed=1, skipped=1)
